requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
apscheduler==3.10.4
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.9
pyjwt==2.8.0
cryptography==42.0.5
asyncpg==0.29.0
aioredis==2.0.1
